creating AzureOpenAIMixin class and have AzureOpenAI and AzureChatOpenAI to inherit from it
